phase3_backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
paho-mqtt>=2.0.0
numpy>=2.1.0
scikit-learn>=1.5.2
joblib>=1.4.2
pydantic>=2.7.1
websockets>=12.0
